extract map logic into mapService and integrate directions plugin preparation
extract map logic into mapService and integrate directions plugin preparation
